refactor(auth): move the account record beside the storage writers (ADR-0033 PR 5.2)
`_auth/account.py` was two modules wearing one name. One half talks to Google
over the network — probe `?authuser=N` for every account a jar can authenticate
as, parse the active user's email out of the page, format the `authuser` value
the wire carries. The other half is *persistence*: read, write, promote and
scrub the `{authuser, email}` record inside `storage_state.json`, the same
document, under the same lock, driven by the same writers that already live in
`_auth/storage.py`.

Keeping those halves apart was not free, and the price was visible in the import
graph: BOTH modules reached into each other from inside function bodies — three
`from . import storage` sites in `account.py`, five `from . import account as
_account` sites in `storage.py`. This commit relocates the record half and
deletes all eight; every one of them is a plain same-module reference now, and
the surviving edge is one-way (`account` -> `storage`) at module scope.
`repair_account_metadata_from_playwright_storage` stays in `account.py` — it
probes, then persists, so it is the one function that legitimately spans both
halves, and it is why the module-scope edge exists at all.

(The governing plan said "8 sites, verified exact" and was right. A mid-flight
recount that reported 4 had missed the parenthesised multi-line
`from . import (\n    account as _account,\n)` spelling.)

PURE RELOCATION. Proven mechanically, not by eye: an AST comparison of all 19
moved functions with docstrings stripped finds exactly three bodies changed, and
each change is the deferred import going away —
`promote_legacy_account` and `write_account_metadata` now call
`update_account_metadata` directly, and `clear_account_metadata` calls
`clear_in_band_account` instead of the one-line `_clear_in_band_account`
delegate that existed only to cross the module boundary. Nothing else in any
body differs by a single node.

The riskiest thing here is `_drain_promotions_at_exit`. PR 5.1's whole finding
was that 38 unit tests passed while real short-lived processes migrated 0 of 6
legacy profiles, because the tests drained the detached worker explicitly and
production never got the chance. That hook is registered with `@atexit.register`
at import time, so moving it modules is exactly the class of change that could
silently return the durable half to 0/6. It cannot here, and the docstring now
says why: the hook is registered by whichever module DEFINES
`read_account_metadata`, and every caller must import that module to reach the
read at all. Verified in real subprocesses rather than trusted: 6/6 durable
migration and 6/6 privacy scrub through the direct import, the `notebooklm.auth`
facade, `get_authuser_for_storage`, and both real CLI commands PR 5.1 measured
(`notebooklm profile list`, `notebooklm auth check`).

Test seams migrate with the bodies. The patches that would have become silent
no-ops — `patch.object(_auth_account, "write_account_metadata")` and its
`clear_account_metadata` sibling in the Playwright-login contract tests, plus
`monkeypatch.setattr(_auth_account, "read_account_metadata", ...)` in the
seam-alias test — now target `_auth.storage`, the module the calls actually
resolve against. Left on `_auth.account`: `enumerate_accounts`,
`_probe_authuser`, `extract_email_from_html`, which did not move.

Facade unchanged by contract: every `notebooklm.auth` name still exists and
still resolves to the very same function object; only the module the alias reads
from moved. `__all__` stays 6.

Guardrail bookkeeping in the same commit:
* module-size ratchet `_auth/storage.py` 2412 -> 3102, annotated as ADR-0033's
  donor-shrinking relocation class with the donor named and measured
  (`_auth/account.py` 980 -> 355, -625). The 65-line gap between +690 and -625
  is prose, itemised in the annotation, in the spirit of PR 4.2's +29.
* `test_storage_writer_boundary`: `_auth/account.py` drops off both write-
  primitive allowlists (it no longer writes anything) and `_auth/storage.py`
  joins the public-`atomic_write_json` importers — `_drop_legacy_account_key`
  targets the SIBLING `context.json`, never `storage_state.json`, so it goes
  through the guarded public primitive exactly as it did in its donor.
* `_auth/refresh.py` is held at its exact 1200 pin: its ADR-0033 raise classes
  do not reach an ordinary re-point, so the import re-point pays for itself by
  tightening the comment block it made stale.

Two comments corrected rather than carried forward, because a deferred-import
comment asserting a cycle that does not exist has already been fixed twice in
this effort. Every surviving function-local `from . import X` on the touched
paths was checked by actually hoisting it and reading the result:
`master_token.py`'s reader import claimed "avoid import cycle" and does not —
hoisting it alone is import-clean, and so is hoisting BOTH directions of the
`storage` <-> `master_token` pair, so the comment now says only what is true
(deferred for symmetry with its two siblings; no cycle either way).
