+++ src/notebooklm/_auth/paths.py
@@ -132,7 +132,7 @@ def _storage_state_lock_path(storage_path: Path) -> Path:
     """Canonical sibling flock file shared by every ``storage_state.json`` writer.
 
     ``save_cookies_to_storage`` (cookie writes) and ``write_account_metadata`` /
-    ``_clear_in_band_account`` (account-metadata writes) all mutate the same
+    ``clear_in_band_account`` (account-metadata writes) all mutate the same
     ``storage_state.json``, so they MUST serialize on the *same* lock file or a
     read-modify-write from one loses the other's update. Deriving the dotted
     ``.storage_state.json.lock`` path here keeps that contract enforced by
