+++ src/notebooklm/_auth/storage.py
@@ -10,6 +10,15 @@ imports in both directions; merging removes those without changing any behaviour
 ADR-0033 PR 4.2 then relocated ``_browser_cookie_filter.py``'s write-time
 cookie-domain filter in as well — misfiled under a ``browser_`` name, but write
 policy applied by three of the writers here — retiring three more lazy imports.
+ADR-0033 PR 5.2 relocated the account-RECORD helpers from ``_auth/account.py``
+(section 7b): reading the record, deriving an in-band-shaped one from a legacy
+sibling, the detached durable-promotion one-shot, and the sibling scrub. They
+are persistence — same document, same lock, driving the writers here directly —
+and splitting them from those writers cost a bidirectional function-local
+import pair (3 sites there, 5 here) which that relocation retires in full.
+``account.py`` keeps the NETWORK identity half (``enumerate_accounts``,
+``_probe_authuser``, page-email extraction, the routing-value formatters) and
+imports this module at module scope; the edge is now one-way.
 
 The file is organised in labelled sections mirroring the former modules:
 
@@ -37,6 +46,13 @@ The file is organised in labelled sections mirroring the former modules:
    lock* as ADR-0029's entry-path-independent guarantee.
 8. **The intent writers** — the seven sanctioned mutations of
    ``storage_state.json`` and its sibling credential files.
+9. **Account records** (labelled ``SECTION 7b`` in the source, where it sits
+   directly after the two in-band account writers it drives) — the record
+   readers, the ``_sanitize_legacy_account_record`` derivation that gives
+   ADR-0029 its anti-wrong-account guarantee, the detached durable-promotion
+   one-shot with its ``atexit`` drain, and :func:`_drop_legacy_account_key`,
+   the sibling ``context.json`` scrub. Relocated from ``_auth/account.py``
+   (ADR-0033 PR 5.2).
 
 This module is the **single sanctioned home** for mutations of
 ``storage_state.json``. It is the only module under :mod:`notebooklm._auth`
@@ -112,6 +128,7 @@ exceptions.
 
 from __future__ import annotations
 
+import atexit
 import contextlib
 import errno
 import json
@@ -130,6 +147,7 @@ from pathlib import Path
 from typing import Any, NamedTuple, Protocol, TypeAlias
 
 import httpx
+from filelock import FileLock
 
 # This module is the SOLE sanctioned user of the module-private
 # ``_atomic_write_json_unchecked`` bypass: the public ``atomic_write_json``
@@ -139,7 +157,13 @@ import httpx
 # ``atomic_write_json``, colliding with the name of the public primitive the
 # guard protects (ADR-0033 decision 2). The boundary is enforced at function
 # granularity in ``tests/_guardrails/test_storage_writer_boundary.py``.
+#
+# The PUBLIC ``atomic_write_json`` is imported alongside it (ADR-0033 PR 5.2):
+# the relocated legacy-account scrub :func:`_drop_legacy_account_key` writes the
+# SIBLING ``context.json``, not ``storage_state.json``, so it must go through the
+# guarded public primitive — the same primitive it used in ``account.py``.
 from .._atomic_io import _atomic_write_json_unchecked as _write_state_unchecked
+from .._atomic_io import atomic_write_json
 
 # ``LockUnavailableError`` is the public, canonical home for the fail-closed
 # lock-failure exception (``notebooklm.exceptions`` — also re-exported on the
@@ -151,7 +175,7 @@ from ..exceptions import LockUnavailableError
 from . import cookie_policy as _cookie_policy
 from . import cookie_semantics as _cookie_semantics
 from . import cookies as _auth_cookies
-from .paths import _storage_state_lock_path, resolve_auth_json_env
+from .paths import _storage_state_lock_path, canonical_storage_key, resolve_auth_json_env
 
 logger = logging.getLogger("notebooklm.auth")
 
@@ -177,18 +201,26 @@ __all__ = [
     "WriteOutcome",
     "WriteStatus",
     "advance_cookie_snapshot_after_save",
+    "clear_account_metadata",
     "clear_in_band_account",
+    "get_account_email_for_storage",
+    "get_authuser_for_storage",
     "in_storage_transaction",
     "merge_cookie_delta",
     "persist_minted_jar",
+    "promote_legacy_account",
     "raise_on_lock_unavailable",
+    "read_account_metadata",
+    "read_account_metadata_from_storage_state",
     "replace_from_login",
     "replace_from_remint",
     "report_on_lock_unavailable",
+    "resolve_account_identity",
     "save_cookies_to_storage",
     "skip_on_lock_unavailable",
     "snapshot_cookie_jar",
     "update_account_metadata",
+    "write_account_metadata",
     "write_master_token",
 ]
 
@@ -1603,6 +1635,8 @@ def merge_cookie_delta(
 
 
 # --- In-band account writers (relocated from ``account.py``) ----------------
+# The WRITE half of the account record; its readers, the promotion one-shot and
+# the sibling scrub follow in section 7b (ADR-0033 PR 5.2).
 
 
 def update_account_metadata(
@@ -1614,13 +1648,15 @@ def update_account_metadata(
 ) -> bool:
     """Persist account metadata atomically inside ``storage_state.json``.
 
-    Relocated from ``account.write_account_metadata`` (the in-band write only —
-    the sibling ``context.json`` cleanup ``_drop_legacy_account_key`` stays in
-    ``account.py``). Full-file RMW intent: **fails closed**, raising
+    Relocated from ``account.write_account_metadata`` (ADR-0033 PR 1.1 — the
+    in-band write only). Its two remaining halves, the facade wrapper
+    :func:`write_account_metadata` and the sibling ``context.json`` cleanup
+    :func:`_drop_legacy_account_key`, joined it here in PR 5.2 and are now
+    same-module calls. Full-file RMW intent: **fails closed**, raising
     :class:`LockUnavailableError` on lock unavailability.
 
     ``only_if_absent`` closes a check-then-act race in
-    :func:`account.promote_legacy_account`: that caller reads the legacy
+    :func:`promote_legacy_account`: that caller reads the legacy
     record and checks whether an in-band record is already present — both
     OUTSIDE this function's lock — before deciding to call this function at
     all. Without a re-check taken under the SAME lock as the write, a
@@ -1635,8 +1671,8 @@ def update_account_metadata(
 
     There is no ``deadline_seconds`` override: every caller takes the standard
     90s full-file-RMW deadline. One used to pass 2s —
-    :func:`account.promote_legacy_account`, back when it ran INSIDE
-    :func:`account.read_account_metadata` and a 90s lock wait would have frozen
+    :func:`promote_legacy_account`, back when it ran INSIDE
+    :func:`read_account_metadata` and a 90s lock wait would have frozen
     an event loop in the middle of what its callers treat as a fast, lock-free
     read. ADR-0033 PR 5.1 moved promotion off that read path onto a detached
     one-shot worker, so nothing is waiting on this acquire any more and
@@ -1647,24 +1683,20 @@ def update_account_metadata(
         and an in-band record was already present under the lock (no-op —
         the caller's stale values were correctly discarded).
     """
-    from . import (
-        account as _account,
-    )  # deferred; no cycle either way (verified)
-
     account_payload: dict[str, Any] = {"authuser": authuser}
     if email:
         account_payload["email"] = email
 
     def _write() -> bool:
-        data = _account._load_storage_state_for_write(storage_path)
-        namespace = data.get(_account._STORAGE_NAMESPACE_KEY)
+        data = _load_storage_state_for_write(storage_path)
+        namespace = data.get(_STORAGE_NAMESPACE_KEY)
         if not isinstance(namespace, dict):
             namespace = {}
-        elif only_if_absent and isinstance(namespace.get(_account._ACCOUNT_CONTEXT_KEY), dict):
+        elif only_if_absent and isinstance(namespace.get(_ACCOUNT_CONTEXT_KEY), dict):
             return False
-        namespace["version"] = _account._STORAGE_NAMESPACE_VERSION
-        namespace[_account._ACCOUNT_CONTEXT_KEY] = account_payload
-        data[_account._STORAGE_NAMESPACE_KEY] = namespace
+        namespace["version"] = _STORAGE_NAMESPACE_VERSION
+        namespace[_ACCOUNT_CONTEXT_KEY] = account_payload
+        data[_STORAGE_NAMESPACE_KEY] = namespace
         _write_state_unchecked(storage_path, data)
         return True
 
@@ -1684,15 +1716,13 @@ def update_account_metadata(
 def clear_in_band_account(storage_path: Path) -> None:
     """Remove the ``notebooklm.account`` key from ``storage_state.json``.
 
-    Relocated from ``account._clear_in_band_account``. Best-effort cleanup:
+    Relocated from ``account._clear_in_band_account`` (ADR-0033 PR 1.1); the
+    one-line delegate that survived in ``account.py`` to reach it was deleted in
+    PR 5.2 when :func:`clear_account_metadata` moved here. Best-effort cleanup:
     swallows lock unavailability and read/parse errors, matching the
     pre-refactor semantics (the reader falls back to the legacy record). No-op if
     the file is missing, unreadable, or carries no in-band record.
     """
-    from . import (
-        account as _account,
-    )  # deferred; no cycle either way (verified)
-
     if not storage_path.exists():
         return
 
@@ -1704,14 +1734,14 @@ def clear_in_band_account(storage_path: Path) -> None:
             return
         if not isinstance(data, dict):
             return
-        namespace = data.get(_account._STORAGE_NAMESPACE_KEY)
-        if not isinstance(namespace, dict) or _account._ACCOUNT_CONTEXT_KEY not in namespace:
+        namespace = data.get(_STORAGE_NAMESPACE_KEY)
+        if not isinstance(namespace, dict) or _ACCOUNT_CONTEXT_KEY not in namespace:
             return
-        del namespace[_account._ACCOUNT_CONTEXT_KEY]
+        del namespace[_ACCOUNT_CONTEXT_KEY]
         if set(namespace.keys()) <= {"version"}:
-            del data[_account._STORAGE_NAMESPACE_KEY]
+            del data[_STORAGE_NAMESPACE_KEY]
         else:
-            data[_account._STORAGE_NAMESPACE_KEY] = namespace
+            data[_STORAGE_NAMESPACE_KEY] = namespace
         _write_state_unchecked(storage_path, data)
 
     # TOLERABLE: the same failure mode the old filelock OSError arm swallowed.
@@ -1726,6 +1756,673 @@ def clear_in_band_account(storage_path: Path) -> None:
             "in-band account clear skipped: storage lock unavailable at %s"
         ),
     )
+
+
+# ==========================================================================
+# SECTION 7b — ACCOUNT RECORDS: READERS, THE LEGACY ONE-SHOT, AND THE SCRUB
+# Relocated from ``_auth/account.py`` (ADR-0033 PR 5.2), which kept the NETWORK
+# identity half. Everything below is *persistence*: it reads and writes the same
+# ``storage_state.json`` document as the writers above, under the same lock, and
+# drives them directly — which is why all EIGHT function-local imports the split
+# used to need (3 there, 5 here) are now plain same-module references.
+# ==========================================================================
+
+
+_ACCOUNT_CONTEXT_KEY = "account"
+
+# The unified atomic profile-state format embeds account metadata
+# inside ``storage_state.json`` under a ``notebooklm`` namespace key, so
+# a single ``atomic_write_json`` covers both cookies and account in one
+# crash-safe commit. ``version`` is bumped only when the in-band schema
+# changes incompatibly — version 1 is the initial shape.
+_STORAGE_NAMESPACE_KEY = "notebooklm"
+_STORAGE_NAMESPACE_VERSION = 1
+
+# --- Detached one-shot legacy-account promotion (ADR-0033 PR 5.1) ------------
+#
+# ``read_account_metadata`` is a READ. It is called per RPC on the token-route
+# path (``refresh._resolve_token_route_kwargs`` -> ``get_authuser_for_storage``),
+# so it must never take the storage WRITE lock. Durable promotion of a
+# pre-v0.5.0 sibling record is therefore fired off the read path as a detached
+# one-shot: the read derives its answer read-only from the legacy record (see
+# :func:`_sanitize_legacy_account_record` — byte-identical to what promotion
+# embeds) and returns immediately, while a background worker does the write.
+#
+# Two pieces of process-global state, both guarded by ONE plain
+# ``threading.Lock``:
+#
+# * ``_PROMOTION_ONCE_PATHS`` — canonical ``storage_path`` strings a promotion
+#   has already been scheduled for in this process. This IS the single flight:
+#   N concurrent reads of one profile schedule exactly ONE promotion, and a
+#   promotion that fails is not retried in-process. Retrying would buy nothing
+#   — the read already returns the right record without it — and would put a
+#   failing write back on a per-RPC path, which is the whole problem. Unbounded
+#   growth is not a concern: real deployments have a handful of profiles.
+# * ``_PROMOTION_THREADS`` — the workers still in flight, so tests can join
+#   them deterministically (:func:`_drain_promotions_for_tests`). Production
+#   never joins; each worker deregisters itself when it finishes.
+#
+# ``_PROMOTION_LOCK`` is a *scheduling* lock, not a storage lock: it is taken
+# only on the legacy-only branch of the read, is held for a set lookup plus a
+# ``Thread.start()``, and is never held across file I/O. The in-band fast path
+# every per-RPC read walks takes NO lock at all (pinned by
+# ``test_auth_account_promotion.py``).
+#
+# Deliberately ``threading``, not ``asyncio``: ``read_account_metadata`` is a
+# synchronous function reached from CLI code with no running event loop as
+# often as from ``async`` code, and the work it defers (``filelock`` acquire +
+# atomic write) is blocking I/O. ``_auth.single_flight`` is the coalescing core
+# for *awaitable* work — it requires a running loop (``asyncio.get_running_loop``
+# in ``_claim``) and would leave the CLI entry path uncovered. Using threads
+# also keeps this module free of lazily-constructed loop-bound primitives (the
+# #1196 class the loop-affinity guard polices).
+_PROMOTION_LOCK = threading.Lock()
+_PROMOTION_ONCE_PATHS: set[str] = set()
+_PROMOTION_THREADS: set[threading.Thread] = set()
+
+
+def _account_context_path(storage_path: Path) -> Path:
+    """Return the context.json path that annotates ``storage_path``.
+
+    Legacy two-file layout: this sibling held ``account`` metadata before
+    the unified format embedded it in ``storage_state.json``. Post-migration,
+    it keeps CLI context state (``notebook_id``, ``conversation_id``) but no
+    longer stores the ``account`` key.
+    """
+    return storage_path.with_name("context.json")
+
+
+def _read_in_band_account(storage_path: Path) -> dict[str, Any]:
+    """Read account metadata from inside ``storage_state.json``.
+
+    Returns ``{}`` when the namespace key is missing, malformed, or the file
+    cannot be read. Callers fall back to the legacy sibling ``context.json``.
+    """
+    if not storage_path.exists():
+        return {}
+    try:
+        data = json.loads(storage_path.read_text(encoding="utf-8"))
+    except (OSError, json.JSONDecodeError) as e:
+        logger.debug("in-band account read failed at %s: %s", storage_path, e)
+        return {}
+    return read_account_metadata_from_storage_state(data)
+
+
+def read_account_metadata_from_storage_state(storage_state: Any) -> dict[str, Any]:
+    """Read in-band account metadata from parsed Playwright storage state."""
+    if not isinstance(storage_state, dict):
+        return {}
+    namespace = storage_state.get(_STORAGE_NAMESPACE_KEY)
+    if not isinstance(namespace, dict):
+        return {}
+    account = namespace.get(_ACCOUNT_CONTEXT_KEY)
+    return account if isinstance(account, dict) else {}
+
+
+def _read_legacy_account(storage_path: Path) -> dict[str, Any]:
+    """Read the pre-v0.5.0 sibling ``context.json`` account record.
+
+    Consumed ONLY by :func:`promote_legacy_account` (the one-shot in-band
+    migration). Never a standing read path — see ``read_account_metadata``.
+    """
+    context_path = _account_context_path(storage_path)
+    if not context_path.exists():
+        return {}
+    try:
+        data = json.loads(context_path.read_text(encoding="utf-8"))
+    except (OSError, json.JSONDecodeError) as e:
+        logger.debug("account metadata read failed at %s: %s", context_path, e)
+        return {}
+    if not isinstance(data, dict):
+        return {}
+    account = data.get(_ACCOUNT_CONTEXT_KEY)
+    return account if isinstance(account, dict) else {}
+
+
+def read_account_metadata(storage_path: Path | None) -> dict[str, Any]:
+    """Read profile account metadata, self-healing a legacy two-file profile.
+
+    **This is a read. It takes no lock and issues no write.** Per-RPC token
+    routing calls it on every request (``_resolve_token_route_kwargs`` ->
+    :func:`get_authuser_for_storage`), so the durable half of the legacy
+    migration is *detached* from it: see :func:`_schedule_legacy_promotion`.
+
+    Unified layout: account metadata lives inside ``storage_state.json``
+    under the ``notebooklm`` namespace key. This reader never returns a raw
+    pass-through of the pre-v0.5.0 sibling ``context.json`` record — a
+    standing read fallback that silently trusted an unmigrated legacy value
+    was the wrong-account hazard #2103's PR-0 closes (a legacy ``authuser``
+    the fallback missed, or a stale one it kept trusting forever, could
+    silently route requests to a different signed-in Google account).
+
+    Instead the three branches are:
+
+    1. **In-band present** — the overwhelming majority of calls, and the only
+       one per-RPC routing walks once any profile has been read: one file read
+       plus a dict lookup, zero locks, zero threads.
+    2. **Nothing anywhere** — ``{}`` (``authuser=0`` downstream), unchanged.
+    3. **Legacy-only** — the record is DERIVED read-only, through the very
+       function promotion itself uses to build what it embeds
+       (:func:`_sanitize_legacy_account_record`), so the caller sees a
+       genuinely in-band-shaped record — never a raw legacy pass-through —
+       whether or not the durable write has happened yet. The durable write is
+       then scheduled once per path, in the background, and the read returns
+       without waiting for it.
+
+    That derivation is the anti-wrong-account contract, and it is what makes
+    the promotion's timing irrelevant to correctness: a promotion that is
+    slow, contended, or permanently failing (read-only profile dir, full disk)
+    changes nothing a caller can observe except how long the sibling
+    ``context.json`` survives on disk. ``tests/unit/test_auth_account_promotion.py``
+    pins field-by-field equality between the derived record and the one a
+    completed promotion leaves behind.
+
+    The ``account`` object records the Google ``authuser`` index used when
+    the profile was authenticated. Profiles from before account-binding
+    shipped (and profiles for users with a single Google account) have no
+    account metadata and use ``authuser=0``.
+
+    Args:
+        storage_path: Path to ``storage_state.json``. ``None`` means the
+            profile is loaded from ``NOTEBOOKLM_AUTH_JSON`` (no sibling to
+            promote from — env-auth profiles skip promotion entirely; the
+            env-auth record is read from the parsed payload by
+            :func:`read_account_metadata_from_storage_state`).
+
+    Returns:
+        Parsed metadata dict, or ``{}`` only when no legacy OR in-band
+        record exists at all.
+    """
+    if storage_path is None:
+        return {}
+    in_band = _read_in_band_account(storage_path)
+    if in_band:
+        return in_band
+    legacy = _read_legacy_account(storage_path)
+    if not legacy:
+        return {}
+    # Re-read in-band before trusting the legacy record. A concurrent fresh
+    # login / account-switch (or this process's own promotion worker) may have
+    # committed one while we were reading the sibling, and in-band ALWAYS wins:
+    # it is the newer, authoritative binding, and preferring a stale legacy
+    # record over it is precisely the wrong-account-routing hazard. Cheap —
+    # this branch is only reached on a not-yet-migrated profile.
+    in_band_after = _read_in_band_account(storage_path)
+    if in_band_after:
+        return in_band_after
+    _schedule_legacy_promotion(storage_path)
+    return _sanitize_legacy_account_record(legacy)
+
+
+def _schedule_legacy_promotion(storage_path: Path) -> threading.Thread | None:
+    """Fire the durable promotion in the background, once per canonical path.
+
+    The caller has already derived its answer read-only, so this exists purely
+    to make the migration *durable* (and to scrub the legacy sibling, a privacy
+    obligation). Nothing downstream of the read depends on it succeeding, or on
+    when it finishes.
+
+    Single-flight: the ``_PROMOTION_ONCE_PATHS`` membership test and the
+    insertion happen under one ``_PROMOTION_LOCK`` hold, so N concurrent
+    readers of the same profile produce exactly ONE worker. It is a one-shot,
+    not a retry loop — a failed promotion is not re-attempted in this process
+    (see the state block above for why).
+
+    ``Thread.start()`` runs INSIDE the lock so a concurrent
+    :func:`_drain_promotions_for_tests` can never observe a worker that is
+    registered but not yet started (``join`` on an unstarted thread raises).
+    ``start()`` returns as soon as the worker is bootstrapped, not when it
+    finishes, so the reader is not made to wait on the write.
+
+    Returns:
+        The worker that was started, or ``None`` when this path had already
+        scheduled one (test/diagnostic affordance; production ignores it).
+    """
+    # Keyed on the CANONICAL path, like every other in-process dedupe in
+    # ``_auth`` (the keepalive throttle, the poke-lock registry, the refresh
+    # flock): two spellings of one file — relative vs absolute, ``~``-prefixed,
+    # or through a symlink — must collapse to one key or the single flight is
+    # silently bypassed.
+    canonical = str(canonical_storage_key(storage_path))
+    with _PROMOTION_LOCK:
+        if canonical in _PROMOTION_ONCE_PATHS:
+            return None
+        _PROMOTION_ONCE_PATHS.add(canonical)
+        worker = threading.Thread(
+            target=_run_promotion_once,
+            args=(storage_path,),
+            name="notebooklm-account-promotion",
+            daemon=True,
+        )
+        _PROMOTION_THREADS.add(worker)
+        worker.start()
+    return worker
+
+
+def _run_promotion_once(storage_path: Path) -> None:
+    """Worker body: promote durably, then deregister.
+
+    :func:`promote_legacy_account` is already best-effort and swallows every
+    realistic failure itself. The broad guard here is for the two things it
+    cannot promise a *detached* caller: an unexpected exception has no caller
+    to surface it to (it would land in ``threading``'s excepthook as a stray
+    traceback), and a daemon worker torn down mid-interpreter-shutdown can
+    raise from arbitrary places.
+    """
+    try:
+        promote_legacy_account(storage_path)
+    except BaseException as e:  # noqa: BLE001 — a detached worker must never escape
+        logger.debug("Background legacy account promotion crashed for %s: %s", storage_path, e)
+    finally:
+        with _PROMOTION_LOCK:
+            _PROMOTION_THREADS.discard(threading.current_thread())
+
+
+def _drain_promotions(timeout: float) -> None:
+    """Join every in-flight promotion worker, each bounded by ``timeout``."""
+    with _PROMOTION_LOCK:
+        workers = list(_PROMOTION_THREADS)
+    for worker in workers:
+        worker.join(timeout)
+
+
+def _drain_promotions_for_tests(timeout: float = 30.0) -> None:
+    """Join every in-flight promotion worker (test/diagnostic helper).
+
+    No production READ waits on this — the whole point of the one-shot is that
+    the durable write never sits on the read path. Tests use it to make the
+    durable half observable, and ``tests/conftest.py`` drains + clears the
+    process-global state between tests so a worker started by one test cannot
+    write into another's ``tmp_path``.
+    """
+    _drain_promotions(timeout)
+
+
+#: Bound on how long interpreter exit will wait for the durable half.
+#: Short by design: a wedged write must not hold a CLI process open, and the
+#: worker stays a daemon so a hung one is still killed if this is skipped.
+_PROMOTION_EXIT_JOIN_SECONDS = 2.0
+
+
+@atexit.register
+def _drain_promotions_at_exit() -> None:
+    """Let the detached promotion land before a short-lived process exits.
+
+    Without this the one-shot is effectively dead in the CLI, which is where
+    legacy profiles actually live. Measured on the first draft: a real
+    ``notebooklm profile list`` against a legacy-only profile migrated it 0
+    times out of 6, and ``auth check`` 1 out of 6 — a pure timing race that
+    real commands lose, because they read the account binding at the very end
+    of their work and the process exits before a daemon worker gets scheduled.
+    The READ was correct every time; what silently never happened was the
+    durable promotion and, with it, the privacy scrub that removes a stale
+    account email from ``context.json`` at rest.
+
+    ``atexit`` handlers run BEFORE daemon threads are torn down, so joining
+    here is what makes the write land. It is bounded and best-effort: a
+    hard-killed process, ``os._exit`` or a fatal signal skips it, and the next
+    read schedules a fresh one-shot — the promotion is idempotent.
+
+    Registration moved modules with the code (ADR-0033 PR 5.2) and that is
+    safe by construction, not by luck: the hook is registered at import time of
+    whichever module DEFINES :func:`read_account_metadata`, and every entry path
+    that can schedule a promotion must first import that module in order to call
+    the read at all. The hazard would be an alias — a shim that re-exports the
+    read from a module the hook does NOT live in — so there is none.
+    """
+    _drain_promotions(_PROMOTION_EXIT_JOIN_SECONDS)
+
+
+def _sanitize_legacy_account_record(legacy: dict[str, Any]) -> dict[str, Any]:
+    """Normalize a raw legacy ``context.json[account]`` dict into the exact
+    record :func:`promote_legacy_account` embeds in-band.
+
+    This is the anti-wrong-account contract's load-bearing piece, and the
+    reason it is ONE function rather than two agreeing implementations:
+    :func:`read_account_metadata` returns this on a legacy-only profile
+    *before* (and, if promotion never lands, instead of) the durable write, so
+    "derived read-only" and "read back after promotion" must be
+    indistinguishable field-for-field. Keeping it shared makes them so by
+    construction; ``tests/unit/test_auth_account_promotion.py`` proves it over
+    a matrix of malformed legacy shapes. Mirrors ``get_authuser_for_storage`` /
+    ``get_account_email_for_storage``'s own sanitization rules."""
+    raw_authuser = legacy.get("authuser")
+    result: dict[str, Any] = {
+        "authuser": raw_authuser if type(raw_authuser) is int and raw_authuser >= 0 else 0
+    }
+    raw_email = legacy.get("email")
+    if isinstance(raw_email, str) and raw_email.strip():
+        result["email"] = raw_email.strip()
+    return result
+
+
+def promote_legacy_account(storage_path: Path) -> bool:
+    """One-shot migration: move a legacy sibling ``account`` record in-band.
+
+    The pre-v0.5.0 two-file layout stored account metadata (``authuser`` /
+    ``email``) in the sibling ``context.json``. This helper promotes that
+    record into ``storage_state.json`` via the canonical storage writer and
+    strips the legacy key.
+
+    **Never called on a read's own thread.** Its three callers are the
+    detached one-shot worker :func:`_run_promotion_once` (scheduled by
+    :func:`read_account_metadata`), the startup layout migration
+    (``migration.py``, which only fires for pre-v0.5.0 two-file profiles), and
+    :func:`replace_from_login`'s ``KEEP_ACCOUNT``-with-no-in-band-record
+    arm, where promoting instead of scrubbing is what stops
+    ``auth import-cookies`` from permanently destroying a legacy profile's only
+    copy of its binding. The read's correctness does not depend on any of them:
+    it derives the same record read-only (:func:`_sanitize_legacy_account_record`)
+    and this function only makes that durable.
+
+    Ordering is crash-safe for the BINDING (never lost), not for the RESIDUE
+    (not guaranteed promptly cleaned up). **Two files, two locks, no shared
+    critical section** — the window is structural, not an oversight:
+
+    * step 1 embeds into ``storage_state.json`` under the dotted storage
+      sentinel ``.storage_state.json.lock`` (:func:`_file_lock`, via
+      :func:`update_account_metadata`);
+    * step 2 strips ``context.json[account]`` under the *separate*
+      ``context.json.lock`` (``filelock.FileLock``, via
+      :func:`_drop_legacy_account_key`).
+
+    Nothing spans both, and nothing may: the two files are locked by two
+    different sentinels held by two different mechanisms, and widening either
+    hold to cover the other step would either invert an existing lock order or
+    unify mechanisms (ADR-0033 plan §1/§5 rules both out). The compensation is
+    therefore *ordering*, not atomicity: embed-then-strip means a crash in
+    between leaves both records present, never neither — the account binding is
+    never lost, which is the correctness property this function exists for. The
+    reverse order would have a window in which the binding exists nowhere.
+
+    What that compensation does NOT buy is prompt cleanup. The NEXT call does
+    NOT reliably take a strip-only branch: :func:`read_account_metadata`'s
+    fast path (``if in_band: return in_band``) returns as soon as in-band is
+    present and never calls this function again (and the one-shot would not
+    schedule a second worker for that path even if it did), so a
+    crash-mid-flight residue can survive indefinitely rather than being cleaned
+    up on the very next read. This is a privacy nicety, not a correctness gap
+    (the authoritative binding is the in-band record, already committed) — it
+    is NOT worth a ``context.json`` existence probe on every read's fast path
+    to close eagerly. A subsequent :func:`write_account_metadata` /
+    :func:`clear_account_metadata` call for the same profile does still strip it
+    (both call :func:`_drop_legacy_account_key` unconditionally).
+
+    When the in-band record already exists for another reason — including a
+    CONCURRENT fresh login/account-switch that won a race against THIS call
+    (see ``only_if_absent`` below) — any stale legacy key IS stripped
+    immediately as residue cleanup (privacy: the old account email must not
+    live on at rest after a re-login); that path runs through this function,
+    not through the fast path above.
+
+    Race-safe: the write is issued with ``only_if_absent=True``, so the
+    decision "is in-band still empty" is made under the SAME lock as the
+    write, not by a separate unlocked check beforehand. Without that, a
+    concurrent fresh login could commit its new record in the gap between
+    this function's own (now-removed) unlocked check and its locked write,
+    and this call would silently overwrite it with the stale legacy values
+    it had already captured — reintroducing the wrong-account hazard this
+    whole migration exists to close, via a race instead of a stale read.
+
+    Best-effort by design — it must degrade to the pre-promotion state rather
+    than raise, both because ``migration.py`` and :func:`replace_from_login`
+    treat it as a completeness step and because its detached worker has no
+    caller to report to. Returns ``True`` only when a legacy record was embedded
+    by THIS call (``False`` both when there was nothing to promote and when a
+    concurrent writer won the race — either way no action was needed from this
+    call).
+
+    Never creates ``storage_state.json``: if it doesn't exist yet, promotion
+    is skipped (returning ``False``) rather than synthesizing a cookie-less
+    file to embed into. Without this guard, a plain READ (``profile list``,
+    ``auth check``) on a legacy profile whose cookies had never been captured
+    (or had been removed, e.g. by the ``COOKIE_VALIDATION_FAILED`` not-exists
+    contract) would itself CREATE a persistent ``storage_state.json`` with no
+    cookies at all — and ``_app/profile.py``'s ``authenticated=storage.exists()``
+    check runs immediately after calling this (transitively, via
+    :func:`read_account_metadata`), so it would flip from correctly reporting
+    "not authenticated" to incorrectly reporting "authenticated" for a profile
+    with zero cookies, purely as a side effect of having been looked at.
+
+    Args:
+        storage_path: Path to ``storage_state.json`` (must be a real file
+            path — env-auth profiles have no sibling and are skipped by the
+            caller).
+    """
+    if not storage_path.exists():
+        return False
+    legacy = _read_legacy_account(storage_path)
+    if not legacy:
+        return False
+    sanitized = _sanitize_legacy_account_record(legacy)
+    try:
+        # only_if_absent=True: the decision "should this write happen" is made
+        # HERE, under the writer's own lock, not by a separate unlocked
+        # _read_in_band_account check beforehand — a check-then-act split
+        # would let a concurrent fresh login/account-switch land in the gap
+        # and then be silently overwritten by these stale legacy values.
+        #
+        # No deadline override: the usual 90s full-file-RMW deadline applies.
+        # It used to be shortened to 2s because this ran INSIDE
+        # read_account_metadata, where a 90s lock wait would freeze an event
+        # loop mid-"read". It no longer does (ADR-0033 PR 5.1) — the caller is
+        # a detached worker with nobody waiting on it, and waiting out real
+        # contention is strictly better than giving up, because the one-shot
+        # never retries in this process.
+        promoted = update_account_metadata(
+            storage_path,
+            authuser=sanitized["authuser"],
+            email=sanitized.get("email"),
+            only_if_absent=True,
+        )
+    except Exception as e:  # noqa: BLE001 — promotion must never raise at its callers
+        # Plain WARNING, no per-path throttle: a persistent cause (read-only
+        # profile dir, full disk) leaves the profile un-migrated, so an
+        # operator needs a default-visible signal rather than one gated behind
+        # -v/--debug. The throttle this replaced existed because promotion ran
+        # on the per-RPC read path and would otherwise have warned twice per
+        # request forever; the one-shot makes that structurally impossible —
+        # a read schedules at most ONE promotion per path per process, so this
+        # branch can fire at most once per path from the read path (plus at
+        # most one each from startup migration and replace_from_login).
+        logger.warning("Legacy account promotion failed for %s: %s", storage_path, e)
+        return False
+    # Reached whether we promoted or lost a race to a concurrent writer —
+    # either way in-band now holds a real record, so the legacy residue is
+    # safe (and, for privacy, necessary) to scrub. _drop_legacy_account_key
+    # already swallows every realistic failure internally (OSError family,
+    # including filelock.Timeout — a TimeoutError/OSError subclass — and
+    # JSONDecodeError), but this call is NOT inside the try/except above, and
+    # read_account_metadata calls this function with no try/except of its
+    # own, trusting it never to raise. Wrap defensively anyway: an embed that
+    # already committed must not be erased by an unexpected exception in an
+    # unrelated cosmetic cleanup step — that would violate this function's
+    # own "never break the read" contract for a reason that has nothing to do
+    # with the embed's success.
+    try:
+        _drop_legacy_account_key(storage_path)
+    except Exception as e:  # noqa: BLE001 — cosmetic cleanup must not undo a committed embed
+        logger.warning("Legacy account context cleanup failed for %s: %s", storage_path, e)
+    if promoted:
+        logger.info("Promoted legacy account metadata in-band for %s", storage_path)
+    return promoted
+
+
+def get_authuser_for_storage(storage_path: Path | None) -> int:
+    """Return the ``authuser`` index recorded for a profile, defaulting to 0.
+
+    Profiles without account metadata (legacy single-account installs and
+    fresh logins that never set an authuser) are treated as ``authuser=0``,
+    preserving existing behavior.
+
+    Returns:
+        Non-negative ``authuser`` index. Malformed values fall back to 0.
+    """
+    raw = read_account_metadata(storage_path).get("authuser")
+    if isinstance(raw, int) and raw >= 0:
+        return raw
+    return 0
+
+
+def get_account_email_for_storage(storage_path: Path | None) -> str | None:
+    """Return the persisted account email for stable routing, if available."""
+    raw = read_account_metadata(storage_path).get("email")
+    if isinstance(raw, str):
+        email = raw.strip()
+        if email:
+            return email
+    return None
+
+
+def resolve_account_identity(
+    *,
+    has_env_auth: bool,
+    storage_path: Path | None = None,
+    env_auth_storage_state: Any = None,
+) -> dict[str, Any]:
+    """Resolve the persisted ``{email, authuser}`` identity for a profile.
+
+    Consolidates a sanitization recipe that used to be duplicated verbatim at
+    ``cli/auth_runtime.py::get_auth_tokens`` and ``_app/auth_check.py::_account_info``
+    (auth cross-boundary ledger shrink, follow-up to #2103): both callers read the
+    in-band account record then apply the identical authuser/email cleanup — an
+    ``int`` authuser clamped to ``>= 0`` (default 0; ``bool`` excluded since it is
+    an ``int`` subclass), and an email stripped-or-``None``.
+
+    The two callers differ only in WHERE the record comes from, not in what they
+    do with it: env-var auth carries no profile directory, so the caller must pass
+    its own already-parsed ``env_auth_storage_state`` (``_app/`` never reads
+    ``os.environ`` directly, and ``cli/auth_runtime.py`` already has the CLI's
+    consolidated ``read_env_auth_json()`` payload in hand by the time it gets
+    here); file-based auth resolves straight from ``storage_path`` via
+    :func:`read_account_metadata`.
+    """
+    if has_env_auth:
+        meta = read_account_metadata_from_storage_state(env_auth_storage_state)
+    else:
+        meta = read_account_metadata(storage_path)
+    raw_email = meta.get("email")
+    email = raw_email.strip() if isinstance(raw_email, str) else ""
+    raw_authuser = meta.get("authuser")
+    authuser = raw_authuser if type(raw_authuser) is int and raw_authuser >= 0 else 0
+    return {"email": email or None, "authuser": authuser}
+
+
+def _drop_legacy_account_key(storage_path: Path) -> None:
+    """Scrub the legacy ``account`` key from the sibling ``context.json``.
+
+    Preserves all other CLI context state (``notebook_id``,
+    ``conversation_id``, …). Best-effort: a failure here does not abort the
+    in-band write. Since the legacy READ path was removed (the reader is
+    in-band-only; :func:`promote_legacy_account` owns the migration), this
+    survives purely as a privacy scrub — a stale legacy key would leave the
+    account email at rest forever with no reader and no writer to remove it.
+    Called by :func:`write_account_metadata` / :func:`clear_account_metadata` /
+    :func:`promote_legacy_account` / :func:`replace_from_login` (the CLI login
+    writer, after its own atomic write).
+
+    This is the one writer in this module that does NOT touch
+    ``storage_state.json``: it writes the SIBLING ``context.json``, so it goes
+    through the guarded public ``atomic_write_json`` and the sibling
+    ``context.json.lock`` (``filelock``) — never the storage sentinel, never the
+    ``_write_state_unchecked`` bypass. Its lock mechanism is deliberately NOT
+    unified with :func:`_file_lock` (ADR-0033 plan §5: that is a cross-version
+    interop change, explicitly deferred).
+    """
+    context_path = _account_context_path(storage_path)
+    if not context_path.exists():
+        return
+    lock_path = context_path.with_suffix(context_path.suffix + ".lock")
+    try:
+        with FileLock(str(lock_path), timeout=10.0):
+            if not context_path.exists():
+                return
+            try:
+                data = json.loads(context_path.read_text(encoding="utf-8"))
+            except (OSError, json.JSONDecodeError) as e:
+                logger.debug("legacy account-key cleanup skipped at %s: %s", context_path, e)
+                return
+            if not isinstance(data, dict) or _ACCOUNT_CONTEXT_KEY not in data:
+                return
+            del data[_ACCOUNT_CONTEXT_KEY]
+            if data:
+                atomic_write_json(context_path, data)
+            else:
+                context_path.unlink()
+    except OSError as e:
+        # Best-effort migration; the in-band reader wins.
+        logger.debug("legacy account-key cleanup failed at %s: %s", context_path, e)
+
+
+def write_account_metadata(storage_path: Path, *, authuser: int, email: str | None = None) -> None:
+    """Persist account metadata atomically inside ``storage_state.json``.
+
+    The account record lands under the ``notebooklm`` namespace key so the
+    (cookies, account) pair commits together via a single
+    :func:`atomic_write_json`. An external reader observing the file
+    mid-update sees either the fully-old or fully-new commit — never a mix.
+
+    The legacy sibling ``context.json[account]`` is best-effort cleaned up
+    after the in-band write succeeds. CLI context state in the same file
+    (``notebook_id`` / ``conversation_id``) is preserved.
+
+    This is the ``notebooklm.auth``-exported facade symbol; it keeps its
+    raise-on-lock-failure semantics (:func:`update_account_metadata` raises
+    :class:`LockUnavailableError` — the documented replacement for the former
+    ``filelock.Timeout``).
+
+    Args:
+        storage_path: Path to ``storage_state.json``. The file is created
+            with empty ``cookies`` / ``origins`` arrays if missing — matching
+            the previous semantics of "writing account metadata never fails
+            because cookies haven't been written yet."
+        authuser: ``authuser`` index used when extracting cookies for this
+            profile (0 for the default account).
+        email: Optional account email to record alongside the index.
+    """
+    update_account_metadata(storage_path, authuser=authuser, email=email)
+
+    # Best-effort: drop the legacy account key from sibling context.json so
+    # the next reader doesn't see the same data in two places.
+    _drop_legacy_account_key(storage_path)
+
+
+def _load_storage_state_for_write(storage_path: Path) -> dict[str, Any]:
+    """Read ``storage_state.json`` for a read-modify-write under the lock.
+
+    Returns a synthetic empty document if the file is missing — matches
+    the earlier behavior where account writes never failed just because the
+    cookie file hadn't been written yet. Corruption is fatal because the
+    primary cookie data can't be recovered from account metadata; surface
+    a ``RuntimeError`` so the caller can prompt the user to re-run login.
+    """
+    if not storage_path.exists():
+        return {"cookies": [], "origins": []}
+    try:
+        loaded = json.loads(storage_path.read_text(encoding="utf-8"))
+    except json.JSONDecodeError as e:
+        raise RuntimeError(f"storage state at {storage_path} is corrupted: {e}") from e
+    if not isinstance(loaded, dict):
+        raise RuntimeError(
+            f"storage state at {storage_path} has unexpected shape: {type(loaded).__name__}"
+        )
+    return loaded
+
+
+def clear_account_metadata(storage_path: Path | None) -> None:
+    """Remove account metadata from both in-band and legacy locations.
+
+    Two documents, two locks, in that order: the in-band record under the
+    canonical storage lock (:func:`clear_in_band_account`), then the legacy
+    sibling under its own ``context.json.lock``
+    (:func:`_drop_legacy_account_key`).
+    """
+    if storage_path is None:
+        return
+    # 1. Strip the in-band record from ``storage_state.json``.
+    clear_in_band_account(storage_path)
+    # 2. Strip the legacy sibling record too (back-compat with old installs).
+    _drop_legacy_account_key(storage_path)
 
 
 # --- Write-time cookie-domain filter (relocated from ``_browser_cookie_filter.py``) ---
@@ -2023,9 +2720,6 @@ def replace_from_remint(
         :class:`WriteOutcome` — ``ok`` on success, ``lock_unavailable`` if the
         bounded storage-lock acquire timed out / the lock infra failed.
     """
-    from . import (
-        account as _account,
-    )  # deferred; no cycle either way (verified)
 
     def _replace() -> WriteOutcome:
         # Carry the existing account namespace BEFORE overwriting (read under the
@@ -2037,7 +2731,7 @@ def replace_from_remint(
             except (OSError, json.JSONDecodeError):
                 existing = None
             if isinstance(existing, dict):
-                namespace = existing.get(_account._STORAGE_NAMESPACE_KEY)
+                namespace = existing.get(_STORAGE_NAMESPACE_KEY)
                 if isinstance(namespace, dict):
                     carried_namespace = namespace
 
@@ -2049,7 +2743,7 @@ def replace_from_remint(
             dict(captured_state), include_domains=include_domains
         )
         if carried_namespace is not None:
-            filtered[_account._STORAGE_NAMESPACE_KEY] = carried_namespace
+            filtered[_STORAGE_NAMESPACE_KEY] = carried_namespace
         _write_state_unchecked(path, filtered)
         return WriteOutcome(WriteStatus.OK)
 
@@ -2147,9 +2841,6 @@ def replace_from_login(
         :class:`LoginWriteOutcome`.
     """
     del io_policy  # reserved; see docstring
-    from . import (
-        account as _account,
-    )  # deferred; no cycle either way (verified)
     from .cookie_policy import (  # noqa: PLC0415 (deferred; true leaf, no cycle either way)
         MINIMUM_REQUIRED_COOKIES,
         cookie_names_from_storage,
@@ -2190,22 +2881,22 @@ def replace_from_login(
         # (3) + (4) Build the ``notebooklm`` namespace (account + opt-ins).
         namespace = {}
         if account is KEEP_ACCOUNT:
-            existing_ns = state.get(_account._STORAGE_NAMESPACE_KEY)
+            existing_ns = state.get(_STORAGE_NAMESPACE_KEY)
             if isinstance(existing_ns, dict):
                 namespace = dict(existing_ns)
         elif isinstance(account, AccountRecord):
             payload: dict[str, Any] = {"authuser": account.authuser}
             if account.email:
                 payload["email"] = account.email
-            namespace[_account._ACCOUNT_CONTEXT_KEY] = payload
+            namespace[_ACCOUNT_CONTEXT_KEY] = payload
         # CLEAR_ACCOUNT: leave the account key absent.
         if include_domains:
             namespace["include_domains"] = sorted(include_domains)
         if include_optional:
             namespace["include_optional"] = True
         if namespace:
-            namespace.setdefault("version", _account._STORAGE_NAMESPACE_VERSION)
-            filtered[_account._STORAGE_NAMESPACE_KEY] = namespace
+            namespace.setdefault("version", _STORAGE_NAMESPACE_VERSION)
+            filtered[_STORAGE_NAMESPACE_KEY] = namespace
 
         # (5) Import backup, inside the lock, before overwriting.
         backup_path: Path | None = None
@@ -2257,10 +2948,10 @@ def replace_from_login(
     # must still scrub directly: CLEAR_ACCOUNT means the caller deliberately
     # wants no account bound, and promoting there would resurrect a binding
     # that was just intentionally cleared.
-    if account is KEEP_ACCOUNT and _account._ACCOUNT_CONTEXT_KEY not in namespace:
-        _account.promote_legacy_account(path)
+    if account is KEEP_ACCOUNT and _ACCOUNT_CONTEXT_KEY not in namespace:
+        promote_legacy_account(path)
     else:
-        _account._drop_legacy_account_key(path)
+        _drop_legacy_account_key(path)
     return outcome
 
 
@@ -2316,7 +3007,6 @@ def persist_minted_jar(
     the default: minting into an existing, unrecorded-owner profile is
     exactly the ambiguous case worth refusing without an explicit ``force``.
     """
-    from . import account as _account  # noqa: PLC0415 (deferred; no cycle either way (verified))
     from . import (
         master_token as _master_token,
     )  # deferred; no cycle either way (verified)
@@ -2333,7 +3023,7 @@ def persist_minted_jar(
         if existed and force:
             logger.debug("persist_minted_jar: force=True bypasses the account-ownership guard.")
         elif existed:
-            existing_owner = _account.read_account_metadata_from_storage_state(data).get("email")
+            existing_owner = read_account_metadata_from_storage_state(data).get("email")
             existing_owner = existing_owner.strip() if isinstance(existing_owner, str) else None
             if not existing_owner:
                 if refuse_unknown_owner:
