+++ tests/conftest.py
@@ -122,7 +122,7 @@ def _reset_poke_state():
        run), so we clear it eagerly to keep tests independent.
     3. ``_SECONDARY_BINDING_WARNED`` — one-shot flag for the Tier 2 cookie
        warning. Reset so tests can independently observe the warning fire.
-    4. ``account._PROMOTION_ONCE_PATHS`` / ``_PROMOTION_THREADS`` — the
+    4. ``storage._PROMOTION_ONCE_PATHS`` / ``_PROMOTION_THREADS`` — the
        detached one-shot legacy-account promotion (ADR-0033 PR 5.1). A read of
        a legacy-only profile schedules a background writer, so teardown must
        JOIN it before clearing: a worker still running when the next test
@@ -133,7 +133,6 @@ def _reset_poke_state():
        half deterministic rather than a race the suite usually wins).
     """
     from notebooklm import auth as _auth
-    from notebooklm._auth import account as _auth_account
     from notebooklm._auth import cookie_policy as _cookie_policy
     from notebooklm._auth import storage as _auth_storage
 
@@ -152,7 +151,7 @@ def _reset_poke_state():
     _auth._POKE_LOCKS_BY_LOOP.clear()
     _cookie_policy._SECONDARY_BINDING_WARNED = False
     _auth_storage._FLOCK_UNAVAILABLE_WARNED = False
-    _auth_account._PROMOTION_ONCE_PATHS.clear()
+    _auth_storage._PROMOTION_ONCE_PATHS.clear()
     yield
     _auth._LAST_POKE_ATTEMPT_MONOTONIC.clear()
     _auth._POKE_LOCKS_BY_LOOP.clear()
@@ -160,8 +159,8 @@ def _reset_poke_state():
     _auth_storage._FLOCK_UNAVAILABLE_WARNED = False
     # Join first, then clear — clearing while a worker is mid-write would let
     # it land in the next test's world.
-    _auth_account._drain_promotions_for_tests()
-    _auth_account._PROMOTION_ONCE_PATHS.clear()
+    _auth_storage._drain_promotions_for_tests()
+    _auth_storage._PROMOTION_ONCE_PATHS.clear()
 
 
 @pytest.fixture(autouse=True)
